Updating Rosenbrock/OPT++ CG tests based on historical review and changing a couple of settings.
This DIFF was originally introduced between DAKOTA 3.3 and 4.0.  However running them side by side at Mike's suggestion revealed small numerical differences accumulating over time.  

Before this commit, the DIFFs showed higher eval counts and a worse solution.  I decided to update the method settings to converge to a reasonable Rosenbrock solution, instead of having early false convergence (as CG is not a good algorithm for Rosenbrock).  This increases the fn evals, but results in a baseline we can consider more of a verification test.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/core/trunk@162 aa0779c3-6ec2-4b84-a212-a405c134cd8d