Bugfix for latest symlink issue reporting by Brian Adams: >>>>> ... the symlink in the workdir.* pointed to the same directory, but the true location of the file, e.g.,     my_sim.inp.template -> my_sim.inp.template instead of     my_sim.inp.template -> ../my_sim.inp.template <<<<<
I should note that the relative path to file (..) is a bit
misleading, since the Symlink logic (wrapper over the system
'symlink' func) does not operate on relative paths.  Instead,
it attempts to determine the full path to the original,
"from", file when creating the symlink.

While this "bugfix" seems to help the situation, I am not convinced
that I understand the logic in the code well enough to "certify"
that symbolic linking is reliable enough for users in production.
Therefore, I'll strongly encourage the use of the "copy" and 
"replace" keywords, when practical, in the documentation
(probably won't get to that till tomorrow).

NOTE:  This change should NOT impact any of our regression tests,
so I expect that we'll still show stable builds in our Nightlies.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/core/trunk@232 aa0779c3-6ec2-4b84-a212-a405c134cd8d