Added EXPORT commands to install targets.
This is necessary to get Dakota to build with CMake within Trilinos.  For
standalone builds, the export target is Dakota-exports (seems consistent with
some of the packages), and it is Trilinos when building within Trilinos.  I
had to modify the install target for many of the packages as well to provide
a similar export command.  Some of these already had export targets, and I
added (hopefully correct) logic to use the Dakota one if provided (which should
always be the case when they are built within Dakota).


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/core/trunk@394 aa0779c3-6ec2-4b84-a212-a405c134cd8d