Track Mike's change to a response map by using map access instead of array access in PSUADE and VBD.  These resolve nightly failures.
In general it won't be easy to find all possible translation errors, e.g., use of [i] should be reconsidered when moving to a map as will insert if not present, default constructing.  I'm doing a quick scan now, but will surely miss some.

Mike: please review.  In particular, do we want the error traps in the event that NaN and Inf get filtered and the returned samples are smaller than asked for?  We could just iterate over the map, not checking its size.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/core/trunk@510 aa0779c3-6ec2-4b84-a212-a405c134cd8d