Invoke WorkdirHelper::which(driver) from the AnalysisCode constructor. The which() member function is an attempt to replace the old, not_executable logic with a Boost.Filesystem-based implementation.
I originally enabled this new functionality in a previous commit (Rev 693), but
then disabled it (Rev 694) to ease Boost Filesystem VERSION 2 stabilization 
and debugging of unrelated aspects of $PATH management, especially the parallel
case in which the dakota_input_file was NULL.

The workdir test suite is working as expected (on Linux) now.
ToDo:  More testing and verify that Trac ticket #4051 can be closed-out.

Request review from Brian Adams.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/core/trunk@737 aa0779c3-6ec2-4b84-a212-a405c134cd8d