Implemented a workaround for otool and Boost .dylib files.  For some reason Boost seems to be a "special case" (at least on ictinus) and otool does NOT provide the absolute path to the libboost_*.dylib files.
So the CMakeCache.txt is probed and the full path to boost libs is used during
the install/packaging step.

I'm not very enthusiastic about this approach, and will consider better ways
to create the deployment tarballs on Mac AFTER the release.

Request review by Brian and/or Dena.

Trac ticket #4058.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/core/trunk@1514 aa0779c3-6ec2-4b84-a212-a405c134cd8d