Bug fixes for AMPL interface revealed by solving the AMPL "diet1" problem as generated by Pyomo:
 * Major: fix use of algebraic mappings and response mapping for case where only algebraic functions are active and the responses are ordered differently between Dakota and AMPL.  While I saw a number of possible fixes, decided to just always call response_mapping from map when algebraic functions are active, not just when core mappings are active.

 * Fix read of AMPL .col and .row files to use getline instead of stream extraction as AMPL labels can have spaces in them.  

 * Minor: Change size_t to int silencing compiler warning (AMPL uses int anyway)


Design notes:

AMPL responses might not be ordered the same way as Dakota responses.  String descriptors are used to properly match them.  In addition, the size of the algebraic response may be less than or equal to the core Dakota response.  

When coreMapping and algebraicMapping both active the core and algebraic responses are gathered up, and then response_mapping overlays the algebraic response, accounting for any potential reordering of the responses to form total_response.

When only algebraicMapping is active, algebraic_mapping was directly populating the final total_response, not accounting for any reordering.

Considered several approaches for the algebraic only case: 
 1. New function to swap ordering in the algebraic response after populated (dangerous).
 2. Change algebraic mappings to be correctly ordered for Dakota consumption (no).
 3. Add logic to current response mapping to handle the algebraic only case, keeping algebraic Response size vs. total_response size distinction (yes).




git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/core/trunk@1656 aa0779c3-6ec2-4b84-a212-a405c134cd8d