Fix parallel tests hybrid and multistart.
Revert most of commit 1158, putting cout back instead of Cout.  By using Cout in strategy, the strategy output ended up getting redirected to dakota.out.1-3 iterator output files instead of stdout and messed up the testing harness as the test output was empty.  My last commit to the perl script should help us catch errors like this in the future. 

This is a regression on capability requested to output everything to specified files, but will be better addressed in the Strategy/Parallelism refactor.

Also found a few other pre-existing uses of Cout that were in error but the test script was missing.

Update dakota_hybrid parallel baseline per serial baseline, moving a DIFF to PASS.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/core/trunk@1800 aa0779c3-6ec2-4b84-a212-a405c134cd8d