Changes to allow the user to specify a nugget for the kriging model in Surfpack (e.g. gaussian_process surfpack). There are two mutually exclusive options:  either the user specifies a real value for the nugget with: nugget = value OR the user specifies find_nugget = 1 or 2. Both of these options involve the kriging algorithm finding the optimal nugget.
If find_nugget = 1, kriging will assume that the condition number of rcondR 
is zero and calculate the nugget needed to make the worst case of rcondR not 
ill-conditioned. 

If find_nugget = 2, it means kriging will calculate rcondR 
(which requires a Cholesky factorization). If rcondR says it's not ill-conditioned,
then use the Cholesky factorization we used to find rcondR, otherwise, 
if rcondR says R is ill conditioned then kriging will calculate the nugget needed 
to make the worst case of rcondR=calculated value not ill conditioned. 




git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/core/trunk@1937 aa0779c3-6ec2-4b84-a212-a405c134cd8d