Changes to Method section to help clarify the new iterator structure.
Mike:  I added a paragraph in the Method specification at the end: 
perhaps it should go at the beginning. 
I also added text to the Minimizer section. 
I really would like to move EGO out of the Surrogate-based minimizer section. 
I know it is derived from that class, but in the spec, it is not grouped together 
with surrogate-based global and local:  it is after ncsu_direct and genie_opt_darts. 
I think this would help the readability if that is acceptable. 




git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2271 aa0779c3-6ec2-4b84-a212-a405c134cd8d