Some partitioning logic refinements for the case of processors_per_server overrides, motivated by this morning's discussion with Rich Drake.  The goal is to avoid a bad master partition that results in significant idleness.
Example: available processors = 15 and processors_per_server = 5, but max 
concurrency > 3 and peer-dynamic is not available (due to multiprocessor 
servers).  Previous logic would select a dedicated master with two servers 
with remainder of 4, wasting a significant portion of the allocation.  New 
logic suppresses the dedicated master selection according to a heuristic 
based on increased idleness.


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2341 aa0779c3-6ec2-4b84-a212-a405c134cd8d