Added utility function:  get_pathext() which is a wrapper around std::getenv("PATHEXT")
The function is NOT used by Dakota (that will come in a subsequent commit) so
the compiler is being used as the "portability" tester for this new utility
function.

Also, some minor cleanup in the code.

ToDo:  Make use of the new function in WorkdirHelper.cpp to "extend" the list
of viable filename extensions supported for executable detection AND commit
a "dakota_windows_ext" test .in file.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2453 aa0779c3-6ec2-4b84-a212-a405c134cd8d