Squashing a bug that has caused docs building in the stable Build Flow to fail a few times (I think).
Currently, both latex-ref/keywords.tex and latex-dev/*.tex must exist before latex_modify.perl is run, or it will die. However, it appears that this condition is not guaranteed by the target/dependency tree in docs/CMakeList.txt.

I modified latex_modify.perl to require a command line switch, either --reference or --developers, so that it processes only the .tex files for the target for which it was executed. Also, modified the commands in the CMakeLists.txt to pass these arguments.

I will merge this change to the branch because it seems like a fairly serious bug. If that gives anyone heartburn, please feel free to email for discussion.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2533 aa0779c3-6ec2-4b84-a212-a405c134cd8d