Add Java sources and build process for XML <--> NIDR translators.  Foss authored the original grammar and XML/nspec translators.  I adapted them to perform the reference manual generation process.  It would be straighforward to generate most of the .desc from the reference manual content and retire .desc.  These are based on DART rev 54692 (sync script included in xml2nidr/)
The CMake option DAKOTA_XML_SPEC (default OFF) builds the tools necessary to work with the XML specification.  The target dakota-gen-nspec will always be built; it generates .nspec from .xml.  The target dakota-gen-xml will hopefully be run a limited number of times to get an initial .xml and then retired.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2647 aa0779c3-6ec2-4b84-a212-a405c134cd8d