Added logic to iterate over globbed paths in WorkdirHelper::link_items(...). Also created a test to cover the new logic.
Request review by Brian Adams.  NOTE: the new logic I added may NOT align
with your vision, so please review and suggest improvements.  Further,
(if so inclined) feel free to modify the logic in the WorkdirHelper::link_items
function and (if necessary) the expected results of this new "unit" test.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2727 aa0779c3-6ec2-4b84-a212-a405c134cd8d