Add the capability to translate or scale NORMAL, UNIFORM, and TRIANGULAR distributions from outer loop parameter insertions.  Previous capability allowed updating of a single distribution parameter at a time, and it has been a long-standing TO-DO to support updates of multiple distribution parameters according to a uniform translation or scaling.  This commit partially addresses this task for the three simplest (and two most common) distribution cases.
After discussion with Laura and Brian, decided on an approach where we allow
convenient definitions of location and scale per distribution, rather than
using a mathematically consistent definition (i.e., mean and std deviation) 
that would require users to perform conversions.  These convenient definitions 
are (location,scale) = (mean,stdev) for normal, (center,range) for uniform, 
and (mode,range) for triangular.  In each case, bounds are updated in a manner 
that preserves the (relative) shape of the original distribution.  To access
these new options, the strings "location" or "scale" are used for the secondary
variable mapping targets, in place of previous distribution parameter names.

Supports scrum task on mixed UQ results generation.

Additional completeness tasks to pursue eventually: (1) extend to other 
distribution types and (2) propagate new mappings to transformation Jacobian 
and Hessian routines (this complexity is the reason it was deferred previously).


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2760 aa0779c3-6ec2-4b84-a212-a405c134cd8d