Fix a bug in NIDR post-processing, where we reported a missing analysis_driver even after it was found.  po_which wasn't returning the full path to the driver, so test for regular file failed.  Changed test to empty as which() checks for whether a regular file.
Also fixes a Windows bug where we'd falsely think we found the driver before we actually located the correct extension for it.  Update the *which() utilities to use bfs::path throughout.

Minor change to initialize preferred path to avoid additional call to file utility if not needed.

Update some comments and rationale in the work directory utilities.

Request WJB review.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2870 aa0779c3-6ec2-4b84-a212-a405c134cd8d