Gather analysis driver tokenization from two places into a utility function tokenize_driver.
Add a new function, developed with WJB, using it that will convert an analysis_driver specified via relative path (leading ./ or ../) to an absolute path, replacing the passed programName entry.

Use this in application interface code to translate drivers as needed.

This should faciliate relative paths to analysis drivers when using work directory.

git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2880 aa0779c3-6ec2-4b84-a212-a405c134cd8d