* Added the generated file VERSION that reports the Dakota version and SVN revision. (Release date not yet included.)
* Changed how DakotaBuildInfo.cpp is generated and packaged to fix a "bug."

Formerly, in an SVN checkout, this file was generated in the build tree during configuration. Under Jenkins control, common_build.cmake copied it to the source tree so that it would get packaged. After packaging, common_build.cmake deleted it from the source tree
to keep the source pristine.

When a user built Dakota from an SVN checkout and packaged source herself, common_build.cmake was not (typically) run, and so BuildDakotaInfo.cpp was not installed in the source tree and packaged.

The essence of the fix is appending the location of DakotaBuildInfo.cpp to the special CPACK variable CPACK_SOURCE_INSTALLED_DIRECTORIES to ensure it gets packaged. (This eliminates the need to copy it to and subsequently remove it from the source tree.)

I also relocated the logic to detect an SVN checkout from src/CMakeLists.txt to the top level so that I could make use of it when building VERSION.


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@2901 aa0779c3-6ec2-4b84-a212-a405c134cd8d