Cleanup following from Russell's change to have one tabular read function throw something more helpful than a string.  We are no longer throwing a string for failed tabular or results file reads.  Instead using specific exceptions derived from std::runtime_error, with a base class FileReadException to enable catching both types.  This makes the catches more intentional, instead of the magic of string.
Also cleanup some of the catch-alls with pass through which were previously necessary to read to end of file in some cases.

Will need to do same for int in evaluation failure management, but wanted an incremental commit.  Request help from other developers to review code and test this with asynch evals.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3022 aa0779c3-6ec2-4b84-a212-a405c134cd8d