Directly construct ExperimentData (leaving empty arrays if not needed) from problem database and clean up client code, including pulling from new input spec variables.  All capabilities should now be accessible for scalar data and field data reading in ExperimentData.
Encapsulate diagnostic print statements in ExperimentData.

I merged with Russell's last changes, but this is an interim commit.  He and I will now work to restore the unit tests that broke as a result of my API changes on ExperimentData.  This commit starts to sketch alternate constructor APIs that will help.


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3031 aa0779c3-6ec2-4b84-a212-a405c134cd8d