Adjust Dakota MPI probe to work when MPI_C and MPI_Fortran are not specified.
Since Dakota only requires MPI_CXX, remove the REQUIRED argument to find_package.  This will allow users to specify MPI_CXX_COMPILER to a custom location and not yield a fatal error when MPI_C and MPI_Fortran are not found.  We already have the necessary error trap for MPI_CXX specifically. Thanks to MSE for the bug report.


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3117 aa0779c3-6ec2-4b84-a212-a405c134cd8d