Adjust sizes and initialize fine-grained Interface evaluation counters, labels, and default ASV on first use.
Previously these were sized based on the function labels from the ProblemDB, but this doesn't cover Recasts which resize the primary functions, e.g., for experimental data or multi-objective reduce, nor field data, as there is only one label per field (group).  Considered changing the constructor chain, but after discussion with MSE, this approach seems better encapsulated.  We still provide an API for clients like surrogate models to initialize fine-grained counters.

The defaultASV is still not fully correct for cases where there is a Response transformation and there are mixed derivatives.  I didn't include in this commit as it might be better encapsulated in Model than here.  It doesn't depend on fnLabels, which is a minor improvement.

Bug fixes:
 * Memory leak in LeastSq confidence interval calculations (array delete)
 * MPICH2 --> MPICH_NAME for MPI_DEBUG to fix MPICH-based debug trap



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3160 aa0779c3-6ec2-4b84-a212-a405c134cd8d