Skip CMake probes for graphics on Windows since that capability doesn't work.
By not nesting the probes for X11, Motif, Threads, threads got enabled in JEGA, which was already happening for Unix, but not Windows builds of Dakota (Dakota doesn't use JEGA evaluation concurrency).  Also committed a change to upstream JEGA CMake to not allow threads on Windows as the CMake logic is incomplete.


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3194 aa0779c3-6ec2-4b84-a212-a405c134cd8d