Localized workaround in NestedModel for Teuchos copy ctor bug.
This just works around the problem for known test issues on Windows, so we can assess DIFFs in time for release.  There are likely issues with gradient_view and other places where function_value_view is wrong. I initiated a more extensive refactor of DakotaResponse to no longer return views by value, then bailed when I learned this should be fixed in Trilinos 12.0, which branches April 14.  

git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3214 aa0779c3-6ec2-4b84-a212-a405c134cd8d