Minor code review and cleanup.
Note: I hoped to include activation of the logit transform but was initially 
unsuccessful -- the cantilever_queso test (a bounded case) and two of the 
rosenbrock_queso tests (the unbounded standard normal cases) fail with it 
turned on.  BMA initially indicated improved performance for cantilever with 
logit, so need to sync up with him on that.


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3233 aa0779c3-6ec2-4b84-a212-a405c134cd8d