Add capability to export the acceptance chain in "user-space" from NonDQUESOBayesCalibration::run_chain_with_restarting().  This aggregates across chain restart cycles, but overwrites the file for each posterior_adaptive cycle (each updated emulator model).  It includes the calibrateSigma parameters in their original unscaled form.
In this initial pass, export is tied to verbosity level (>= NORMAL_OUTPUT), 
the file in non-annotated, and the filename is hardwired to 
"dakota_mcmc_tabular.dat".  Will confer with BMA and LPS to see if they 
have other preferences, including elevation of this capability to the 
input spec with filename and annotation option support.

Also some minor cleanup related to DataFitSurrModel export.


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3616 aa0779c3-6ec2-4b84-a212-a405c134cd8d