Initial deployment of DataTransformModel to Bayesian methods.
Ths is a step toward full use of RecastModel in Bayes that:
 * Loads experiment data at base class
 * Encapsulates hyper-parameters and their effect on covariance-scaled residuals in DataTransformModel (eventually in likelihood); with default operations when no hyper-parameters.
 * DREAM and QUESO likelihood routines are now agnostic to hyper-parameters.  MAP to come next.

DREAM is fully migrated.  For now, I'm maintaining residualResponse for QUESO as I make increments of improvement, but will eventually retire it.

Also had to adjust RecastModel lightweight ctor for when variables sizing changes, but map is NULL.

DIFFs are same as current RHEL6 build.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3730 aa0779c3-6ec2-4b84-a212-a405c134cd8d