Bayesian calibration: Update the MAP pre-solve neg_lost_post mapping to use the full likelihood formulation, including hyper-parameters, and all likelihood terms.
DataTransformModel now fully encapsulates transformations of functions, gradients, and Hessians, including effects of Covariance and any estimated hyper-parameters.

Added a new ExperimentData function that tracks the dependency of residuals on hyper-parameters.  May decide to make that the primary way to manage the mappings.

This induces DIFFs in two tests, which can likely be blessed, though I'm hoping to review from Monday's build vs. this one:
  pdakota_cantilever_queso_adapt
  pdakota_10d_rosenbrock_queso



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3756 aa0779c3-6ec2-4b84-a212-a405c134cd8d