Move EfficientSubspaceMethod build process into ActiveSubspaceModel, a specialization of RecastModel.
This lays the foundation for using subspace Models in more general contexts.  However it is an incremental step.  I tried creating a new user-facing subspace model type that could be passed to a method, but there are too many construct time deps on model size.  Construct vs. run-time init will need to be reworked for Iterators.

Along the way:
 * Generalize to handle all active variables (continuous are mapped, discrete carried along)
 * Generalized to properly handle primary and secondary response functions
 * Only compute functions, not gradients, when performing reduced-space UQ

Minor:
 * Fix indexing error when using DataTransformModel with discrete variables
 * Remove uneeded ParallelLibrary header from RecastModel

TODO:
 * Review use of communicators and transformations with MSE/JAM



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3879 aa0779c3-6ec2-4b84-a212-a405c134cd8d