Refactor ActiveSubspaceModel into a user-facing, problem-DB constructed Model that wraps another subModel.
For now, this is an alternate to other models in the spec and requires a pointer; it could be made an option to apply to any other model, which might be cleaner.  Also, have to decide whether to use the dace_iterator paradigm due to the number of method controls, perhaps making the subspace ID process a helper iterator.

Simplifies communicator management.  Moved init comms to normal place in Model, since it knows its requirements for building, though ActiveSubspace still has to specialize.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3897 aa0779c3-6ec2-4b84-a212-a405c134cd8d