Initial working version of KL-based random field model, joint work with LPS.
For the data read case, we can now generate and forward propagte using a KL approximation to a random field.  The RandomField model uncertain variables are augmented with additional random coefficients xi that parameterize the KL model.  When evaluate() is called on the model, a random field is generated, then the submodel is evaluated (and is responsible for accepted the RF data).

More needs to be done to generalize this, but the RFModel can be operated on by an LHS sampler that samples it's uncertain vars + xi.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@3973 aa0779c3-6ec2-4b84-a212-a405c134cd8d