Credible/prediction intervals for Bayesian methods
Initial capability to cache the simulation or emulator function values corresponding to the acceptance chain for QUESO.  Works be either reevaluating a surrogate or looking up in the cache.  This will enable calculation and export of credible and prediction intervals.

The demonstration only computes this in debug output for now until we make it more robust.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@4063 aa0779c3-6ec2-4b84-a212-a405c134cd8d