Push down the sampleType default assignment in order to allow a MC default instead of an LHS default for NonDMultilevelSampling.  This removes the hardwired MC assumption and allows use of LHS within NonDMultilevelSampling. Multilevel LHS, so far, is surprisingly effective even under a MC estimator variance assumption.
Note: the new Wilks sampling option has a dependence on the sampling type in
the NonDSampling ctor, but the default is not reassigned until the derived
class.  For now, Wilks requires an explicit MC override, so no change in
behavior is currently possible, but may want to rethink this design in time.


git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@4089 aa0779c3-6ec2-4b84-a212-a405c134cd8d