Correlation calculation cleanup for DAK-361
In preparation for coming changes to simple and partial correlation calculation, consolidate some common correlation setup code in convenience functions.  Specifically:
 * Determining valid finite samples
 * Converting values to ranks

Added a new function to Variables (could instead be a free function) to retrieve the active vars as a RealVector across all types.  This can also be used in other conversions between Variables and samples.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@4153 aa0779c3-6ec2-4b84-a212-a405c134cd8d