Fixed bug in the Bing Li truncation method that was producing diffs on Red Sky and TLCC2. Updated the test baseline to reflect the correct results.
Below is an explanation of the bug:


The code below is similar to the problem code and will produce undefined results:

RealMatrix A = ....
RealMatrix B = ....

A.multiply(Teuchos::TRANS, Teuchos::NO_TRANS, 1.0, B, A, 0.0);

Instead I changed it to something like:

RealMatrix A = ....
RealMatrix B = ....
RealMatrix A_copy = A;

A.multiply(Teuchos::TRANS, Teuchos::NO_TRANS, 1.0, B, A_copy, 0.0);

which seems to have fixed the bug and diffs. With this change the Bing Li method is producing much better results.

The reason for the diffs is because TLCC2 and Red Sky were both using MKL instead of the default LAPACK and BLAS which gave a different undefined result than the test baseline.



git-svn-id: svn+ssh://software.sandia.gov/svn/dakota/trunk@4215 aa0779c3-6ec2-4b84-a212-a405c134cd8d