Bayesian: validate user-input sample and update specs
Dakota 6.4 has a bug in surrogate-based Bayesian calibration: if a
small number of chain_samples (<100) is specified with an emulator the
default is to perform 0 cycles of size 100.  This change will
configure 1 cycle of length chain_samples.

Also trapped a potential error case where the number of proposal
updates requested is greater than the number of chain samples.