Improved active subspace scaling for multiple QoI
Previously active subspaces were formed over multiple QoI by scaling
the gradient matrix by the mean values of each of the QoIs. After
talking to Paul Constantine, scaling by the mean norm of the gradient
might be better. This is now the default behavior but can be changed
through the input spec. Also added a few tests to exercise this
capability.