Enable automatic detection of parallel launch with aprun
Users reported that Dakota does not correctly detect parallel launch on
Cray platforms, which use the ALPS (Application Level Placement Scheduler)
aprun commmand. They submitted a patch based on detecting an environment
variable, which I confirmed works on mutrino. briadam recommended we add
an override of the automatic detection so that when similar issues arise
in the future, users have a workaround.

The override is an environment variable, DAKOTA_RUN_PARALLEL. If it is
set and begins with 1, t, or T, Dakota will attempt to run in parallel mode.
If it is set and begins with anything else, Dakota will run in serial
mode.