Initial changes to test on Cray Linux Environment
The aprun command is used to MPI launch executables on the compute nodes
on Cray XC systems such as mutrino and trinity. This change to
dakota_test.perl detects when it is being run within a submitted job on
a Cray XC system and sets the global using_aprun to 1. For parallel
testing, aprun is used in a manner similar to the way mpirun/mpiexec
are used on other platforms. For serial tests, the aprun command must be
used with DAKOTA_RUN_PARALLEL set to false to override Dakota's parallel
launch detection logic.