Provide a list of tracker/proxies for usage tracking
Before this commit, only a single usage tracking url/proxy could
be specified at build time, necessitating separate builds of
Dakota for different networks.

This commit permits specification (still using a compiler define)
of multiple tracking url/proxies. Dakota tries them in order.
Those that fail (timeout definition of 2 seconds retained) are
pruned from the list so that on subsequent attempts, the timeout
cost is not incurred again.

This means if tracking is enabled, Dakota runs on some networks
may be lengthened by a couple of seconds as urls are tried and
discarded. Probably for most Dakota studies, that delay is
swamped by the duration of the actual analysis.

The payoff is simpler build/test/deployment: We only need to
build one set of binaries now.