Shorten the format of Dakota revision
The Dakota revision that is printed when Dakota is run and that
is reported for usage tracking is the abbreviated SHA1 for the
most most recent commit and the date and time for the most recent
commit. The date and time is printed in a perhapse overly detailed
or verbose format.

E.g.

Dakota version 6.4+ (stable) released Oct 21 2016.
Repository revision 83c1aaa on Fri Oct 21 16:54:36 2016 -0600 built Oct
21 2016 16:31:26.

This commit shortens up the revision string to:

Dakota version 6.4+ (stable) released Oct 21 2016.
Repository revision 2cb85ea(2016-10-21) built Oct 21 2016 21:48:37.