Turning off debugging (DAK-1088)
Switched RelWithDebugInfo to Release and deleted DAKOTA_DEBUG
everywhere they occur. This includes platforms that we no longer
distribute, for the sake of completeness. The optimization
level is unaffected,, and so hopefully this commit will not result
in any test regressions.

DAKOTA_DEBUG doesn't seem to have any effects, anyway(?)