DAK-1105: Updated improved performance tests.
Added a couple of tests for new initial and minimum mesh size
controls.  Updated baseline to include these new tests and to reflect
the smaller number of function evaluations required to to more
reasonable default stopping criteria.  Did this for NOMAD tests in
both dakota_nomad.in and dakota_scaling.in.