Two minor updates to Bayesian tests.
The first update is the explanation and reference
for the direct analysis_driver called bayes_linear.
The second is the addition of asynchronous evaluations
for the first test of the Bayesian experiment design
tests (#s0 in dakota_bayes_expdesign.in).  This test
now utilizes asynchronous function evaluations for
both the high and low fidelity models.