Docs for optional interface in nested model
In the old (5.4) Reference Manual, we relied on comments in the
Developer's manual to explain this feature, especially the way
responses from an optional interface are combined with results
from the sub-iterator. This commit expands and (hopefully)
clarifies this information, and also makes it a first-class
citizen of the Reference Manual.