NaNs and Infs as corr coeffs (DAK-1238)
Adding a message that provides possible explanations for the
appearance of NaNs and infs as correlation coefficients. Although
NaNs and Infs are appropriate numerical results that indicate
things like insufficient sample size, they are a recent change
and make some users nervous. The new message is intended to
reassure users that Dakota isn't malfunctioning.