Bug DAK-1336 MCMC tabular in standardized space
User reported that when using the standardized_space input file
option, the exported chain points do not contain correct response
values.

This occurred because of multiple model transformations for
ProbabilityTransform and DataFitSurr together with a logic error in
function value recovery.  Chain recovery should now work with various
combinations of emulator/none and standard/native.

A limitation of the current implementation is if the MCMC model is of
type HierarchSurr, it may get re-evaluated at function value recovery
time.  We may have to reconsider the logic depending on what happens
with the eval cache in that case.