Automatically convert variables to int or float
Originally I wanted to keep the variable values as the exact
strings extracted from the parameters file, reasoning that the
user might want them that way and that it was easy to convert
them to floats or ints as desired.

But, it's annoying to keep writing int(foo) or float(foo),
and situations when someone would want somethign that looks
like a real number to be treated as a string instead of a real
number seem few and far between.

So, this small change converts things that look like ints to ints
and things that look like floats to floats. If it becomes a
problem, I can add an option to override it or make the original
strings available.