DAK-1441: Fixed error in user's manual cylinder head example.
Eq. 20.9 had an error in which t_wall was computed based on the
difference between the intake and exhaust diameters. The source
code had the right equation which depends on the sum of the two
diameters. The user's manual was fixed to be consistent with the
code. In a previous version of the Dakota manual (1999), t_wall
depended on the sum of the two, and thus the hypothesis is that
is the right form and the most recent versions had a typo.