DAK-1331: Harden optimization data transfers/adapters for release
Based on Brian's feedback on the data adapters, I'm moving the only
adapter that was a class method to a free function.  This is consistent
with the others.  The previous form was done to avoid having to pass
needed data through the function arguments.  This data is now
obtained from a passed Model object.