DAK-1631: Numerical grads and DataTransformModels
Fix for a bug reported on dakota-users. Numerical gradients
returned to the iterator when using a calibration_data_file
in combination with scaling were incorrect.

The problem was in RecastModel::db_lookup(), called by the
envelope from Model::estimate_derivatives(). The lookup request
was not being forwarded down the RecastModel recursion chain, and
was returning responses that had been unscaled, but not
differenced with calibration data.

Added a test case and baseline to protect the change.