DAK-1708 : Propagate traits to remaining TPL wrappers
Some of the tests needed to ensure consistent behavior when switching
to using adapters were broken.  This commit propagates traits usage to
the LeastSq base class (as was done for Optimizer) so that correct
queries are returned for constraint support in the TPLs.

I used the dakota_scaling test as a guide to enable the traits needed
for all the subtests to pass.  This actually is a nice way to find out
which traits a package supports and is tested against.

In addition, I manually verified consistent behavior with the devel branch
for daktoa_apps, dakota_ncsu_textbook, dakota_mogatest, dakota_cyl_head,
dakota_nonlinear_cg, dakota_optdarts, dakota_herbie_smoothherbie_shubert,
dakota_cantilever and dakota_textbook (but noting that test 9 reported a
bad termination criterion for devel but seemed to work okay for the
traits2 branch).  All of these were chosen based on the testing guidelines
in the table here:

https://sems.sandia.gov/confluence/display/DAK/Optimization+TPL+Interface+and+Testing+Refactor?preview=/3244317/10093152/reg_testing_1708.xlsx

I also verified that the unit tests all pass.

I didn't intend to encroach on a story I don't own but made these changes
of necessity.