Improve error checking in NonDMultilevelSampling
Improved the error checking for solution levels and costs in the ctor of
NonDMultilevelSampling.  Required addition of a new Model virtual function
to reutrn number of solution costs, which defaults to zero (number of solution
levels defaults to one, as there is a nominal simulation config; but there is
no default cost).