Adding multilevel PCE based on a hierarchical surplus approach.
Tracks corresponding Pecos commit.

This approach provides an alternative to the existing PCE routines for
multilevel regression and multifidelity PCE wherein the HierarchSurrModel
is placed in MODEL_DISCREPANCY mode and the approximation levels are distinct.
In this new approach, the approximations are recursive and each level
approximates the difference between the response data and the aggregate
approximation at the previous level.

Anticipated advantages, as annotated in the code comments:
> more consistent with HierarchInterpPolyApproximation,
> greater sample set freedom across levels,
> reduced cost from not requiring matched Q^lm1 observations,
> telescoping consistency of level surrogates,
> local error estimates relative to previous surrogate,
> simplified multilevel data import