Changes to surrogate initialization in NonDBayes
This is a fix for DAK-1768, where chain_samples was not being initialized properly in DREAM when an emulator was
specified.  This fixes a bigger issue:  except for QUESO, users of other Bayesian algorithms typically
want a one-time surrogate which is not adaptive and does not involve adaptive posterior refinement.
We have moved the initialization of the posterior adaptive surrogate algorithm controls to QUESO.
NonDBayes now has the defaults for the derived classes but these are overriden in NonDQUESO.