Start to harden Wasabi, including reading observational data.
Per DAK-1783, these are some changes to harden Wasabi.
Specifically, observational data is now correctly read
from either the specification of means and variances
or from a file.  Multiple data points are now allowed.
If the prior and posterior samples are from the same sample set,
the posterior responses are not re-evaluated from the surrogate.
The output processing has been cleaned up a bit.

TODO:  The observational data needs to account for multiple
calibration terms.  Also, we need to read this directly from
a calibration data file if the user specifies that.  The
output options of generate_posterior_samples and evaluate_posterior_density
need to be revisited in terms of their combinations.
Finally, some checks need to be added so the user knows when the KDE
has problems or when the observational data is outside the
bounds of the model data.