Enable LSQ confidence intervals in presence of scaling
Enable calculation of CIs when variable and/or response scaling is
present.  This required the derivative of the scaled responses with
respect to the original unscaled parameters since the scaled responses
are used in the nonlinear regression, but the user expects results in
the unscaled parameter space.

Implemented as a conditional in ScalingModel::response_modifys2n as a
cleaner separation of variable and response scaling wasn't possible
due to mixed terms in the derivatives.

Updated baselines for scaling-related tests to account for duplicate
evals when computing confidence intervals.

Resolves DAK-1812 as part of DAK-1810