Update the input spec for WASABI.
This commit updates the input spec for WASABI, specifically changing
initial_samples to pushforward_samples and making it a separate
data member.  Also put back the alias of initial_samples for pilot_samples.
Reworked some of the ordering of the keywords for WASABI and updated
the test.  This addresses DAK-1783.