Relax unique id requirement in XML schema
We aren't currently using the id field separately from the name field
(though the GUI currently refers to it in a number of places).  So we
no longer enforce unique keyword ids in the schema.  Eventually we'll
remove the attribute in dakota.xml.

Corrected the 4 outstanding errors that occur when validating
dakota.xml against dakota.xsd.