Bug fix: Model::db_lookup was retrieving garbage
Fix a latent bug, shown with scaling + data transformations where
garbage gradients might get retrieved in the course of numerical
Hessian computation.  Issue was transformation and passing of
ActiveSet to the subModel.

Also, my last commit introduced a but in confidence intervals missed
by a problem with dakota_diff.perl.

Updated baselines, accepted a regression in a test for now (that did
better with a garbage/identity Hessian than the right one).