Constrain sample profile in RIP_SAMPLING algorithm to retain shape while enforcing upper bound on collocation ratio.
THe RIP_SAMPLING multilevel algorithm shapes the sample profile based on the Restricted Isometry Property (RIP).
This upper bound is not sharp, and the resulting sample allocations are too large.  Moreover, the sparsity
estimates tend to grow for compressible QoI as increased samples drive increased accuracy in less important
terms --> cross validation scores improve for dense solutions, which defeats the point.  To control this effect,
we retain the shape of the RIP profile, but enforce an upper bound on the collocation ratio (currently 2.) that
is applied against a non-adapted candidate basis order.