Added a section to the Theory manual on the measure-theoretic approach for inference that has been implemented with WASABI.
This section on consistent Bayes was written by Tim Wildey.  As part of this addition,
we generalized the introduction to the Bayesian chapter in the Theory manual
to include all of the topics now covered, including model discrepancy,
Bayesian experimental design, information theoretic tools, and
the measure-theoretic stochastic inversion approach.