DAK-1747:  Fixed rol_ls entry in dakota.xml.
The </keyword> identifier for ROL-related keywords was in the wrong
place, causing the build to barf when ENABLE_SPEC_MAINT is turned on.
In particular, the rol_ls entry was messed up in dakota.input.nspec.
The version of dakota.input.nspec in the repository seems to be
correct, so no new commit of that file.