Bayes Hi2Lo: Enable batch evals of HiFi Model
Add a static function on Model that accepts a matrix of samples to
evaluate and returns a matrix with the corresponding function value
data.  It manages synch vs. asynch under the hood for the client.

Update Hi2Lo / mutual information calibration code to use this
component.

Completes DAK-1878 in support of DAK-1794.  Compiles and does not
regress any tests.