DAK-1880: Fix rpath issues on Linux
Running dakota 6.7 on Linux gave a missing library error.
Specifically, libquadmath.so.0, which we install to /bin,
could not be found. Examining the output of LD_DEBUG=all
shows that libdakota_fortran_src.so is the first
requestor of this library.

The resolution was adding $ORIGIN/../bin to RPATH.

A further issue that remains unresolved is bad or missing
RPATH information in our installed (not built) libs. The first
culprit is libgfortran.so.3, which contains an absolute path to
a sems folder. Currently, it isn't hurting us because of the
order in which ld.so finds and loads requested shared libs*, but
it conceivably could some day.

(* By running dakota with LD_DEBUG=all, I found that ld.so loads
libraries and their dependencies breadth-first. The second library
loaded when dakota is run is libdakota_src_fortran.so. It requests
libgfortran.so.3 and libquadmath.so.0, which are now discovered
in bin/ because of the RPATH fix in this commit.

libquadmath.so.0 is also a dependency of libgfortran.so.3 itself.
Because libquadmath.so.0 was already discovered, the dependency is
satisfied without further searching. However, if we later find ourselves
in a state where that's not true, we'd be back to a missing library
error because of the bad RPATH in libgfortran.so.3.)

Another, more remote area of concern is the Boost libraries, which
contain no RPATH information at all. There probably are other
examples.

We could modify the RPATHs, which we do on Mac using the universally
available utility install_name_tool. A similar utility called chrpath
is available on Linux, but unfortunately it is much less capable. It
can't add an RPATH to an application or lib, only modify an existing
one. And the RPATH can't be longer than the current one. Another tool
called patchelf exists for Linux, but it is not yet widely available.