DAK-1949: Migrated Boost smart pointers to Standard Cxx11 equivalents
Boost serialization in the 1.53 lacks ability to serialize std::shared_ptr.
Boost 1.56 is the first version that supports serialization of std::shared_ptr
and it will be a while before Dakota can migrate to Boost 1.56.

Therefore, I added an "adapter" to convert back-and-forth between the two:
  to_boost_ptr(const std::shared_ptr<T>&) and
  to_std_ptr(const boost::shared_ptr<T>&)

Not excited about this, so request a review from Brian Adams.