DAK-1931: Restore previous sharper unit testing
Something is still not quite right with the gradient information, and
this is causing the unit tests to take very long times to reach a less
converged state than previously.  I'm continuing to look into the the
issue with th grdients but want to preserve efficient and sharp (tighter
tolerance) unit testing while making code changes.