DAK-1031: Refine ROL update/evaluate
After some trial and error and subsequent discussion with Brian, it turns
out that the attempt to further imrpove efficiency by skipping the call to
evaluate the model based on the incoming variables state is complicated by
the need to ensure consistency between the model variables and responses,
especially at the beginning of the optimization.  For now, it seems best
just to go ahead and call evaluate and rely on the evaluation database
to provide duplicates when available.