DAK-1979: Sketched out Hessian addition for ROL objective
Did some pair programming with Russell to sketch out what's needed to
add Hessian support in the ROL wrapper.  Only did this for the
objective because the scope of the story is bound-constrained
problems.  There will need to be follow-on activitie under the
constrained stories to capture the needs there as well.  Created a new
branch off of rol_integ to avoid breaking and therefore interfering
with anything in flight there.