DAK-1979: Conditional solver settings for Hessian support
Added all (I think) of the conditional logic needed to instantiate,
set, and invoke ROL use of a Dakota/user-provided Hessian.  Unit tests
now run cleanly.  Also added a unit test that is a variant of the
bound-constrained test with analytic Hessians and confirmed that our
hessVec is being called.  The test runs to completion, though not sure
what the solver is doing since we haven't yet actually implemented the
hessVec product.  Our invHessVec is not being called, but as far as I
can tell, that's only because the particular solver we have defaulted
to for bound-constrained problems does not use it.  Will double check
to see if the other default solvers use it; may be able to eliminate
it if not need by the set we're planning to deploy.