DAK-1979: Code clean-up and documentation.
More code re-organization in the ROL wrapper, mostly in the cpp file
to follow the conventions already captured in the hpp file.  Added
more documentation about what the code is doing.  There are a number
of "QUESTION" comments that indicate where either clarification
documentation is needed or where sanity checks on code are needed.
"CLEAN-UP" comments are of two types, one of which is identifying
apparent litter and the other is identifying where we may want to
update adapters later on.