DAK-1979: Added bound-constrained regression tests
Added regression tests to cover challenging responses, different
combinations of analytic and approximated derivatives, and larger
numbers of variables.  The underlying assumption is that these span
scenarios for which we would want to catch behavior changes in nightly
tests.  Set the stopping criteria for these tests to be as loose as
possible but still allow convergence to the solutions.  Did not try to
cover multiple combinations of method parameters, as that is well
covered in unit/verification/performance tests.