DAK-2007: ROL Wrapper: Assess and correct advanced use cases
 - revised ROL re-entrant API

Instead of providing a ROL-specific reset_problem(...) method, I now
rely on the Iterator::iteratedModel contents to set initial values,
constraints, bounds, etc. in core_run() as is done with other TPLs.
I kept the ROL-specific call to reset ROL solver options via the Teuchos
ParameterList mechanism.  The unit test shows how this is done so as
to preserve previous behavior, and the dakota_hybrid test now reflects
better initial values obtained from the previous optimizers in the
sequence when starting ROL.

Note that this change precludes the ability to reset an existing ROL
solver in an attempt to preserve internal solver state (as was the intent
of the previous API), but that capability seems to be a topic for
after the May release.