DAK-2068: Incorporate feedback from Bill
These changes should help with DAK-1997.

For some reason, the new baseline in dakota_rbdo_cantilever_trsb2.in
differs from my results when using the RHEL7 Docker container.  This
suggests something is not completely consistent between Docker and the
RHEL7 Jenkins test environment.  Or maybe this test is inherently noisy...