Temporary patch to unstick docs
Some of the info in the User's Manual is dynamically
generated by running dprepro. This has been broken by
updates to dprepro, which included preserving the old
dprepro by renaming it to dprepro.perl. This commit
causes dprepro.perl to be run to generate info for the
manual, which hopefully will get docs building going
again.  As we update the User's Manual with info
about Python dprpero before the docs freeze, I'll
revisit this.