Close DS datasets after they are used
Examination using h5py showed that the dimension scales weren't
being attached properly. The problem appears to be that the DS
datasets were being closed before they were attached, which I
think invalidates their IDs. (The ret_code for the attachment
operation was != 0.) I moved the closes to a point after they
are attached, and things are working now.