Clean up some pointers and add comments
This is helpful reading about identifiers.

https://support.hdfgroup.org/HDF5/doc/Advanced/UsingIdentifiers/index.html

I initially made the error of assuming that every HDF5 object had a
unique ID. The truth is, an object can have many IDs, and they all
have to be closed. I think the information at this link also implies
that it's okay to close the identifier for, e.g., a dataset's
dataspace before closing the dataset. The C++ API docs also
seem to suggest that an object's destructor closes its ID. So, if
an object is created on the stack (that is, not with new), its ID
will be closed when it goes out of scope.