DAK-2315:  Proof of concept for 2D dataset with multiple dimension scales
Implemented Adam's design for storing moments and bounds for confidence intervals in a 2x2 matrix that has two string-based dimension scales attached (one for each axis of the matrix).  This functionality is also (partially) covered with some test asserts.  Still need to figure out how to harden this code for more general use.