DAK-2315: Use hyperslabs to write subselections
Added a demonstration of writing a portion of a dataset
using hyperslab selections. This capabiliity will be needed to
write non-contiguous data in memory, such as vectors of vectors,
to datasets. Using dataspaces to perform "partial I/O" is
described in great and sometimes confusing detail in chapter 7
of the User's Guide.

I also replaced the C API query for the existence of /methods
with a call to .exists().