DAK-2345: Test the moments and dimension scales
On my RHEL6 machine, it is currently necessary to build
with an unusual set of SEMS modules for the test to work.

In addition to the typical stuff like CMake and Boost:

First, load sems-hdf5/1.10.2.

Then make sure that sems-gcc/4.8.4, sems-python/2.7.9,
and sems-openmpi/1.8.7 are loaded. Usually I build with
the system OpenMPI module, but sems-h5py requires a SEMS
openmpi.

Lastly, load sems-h5py. A warning will appear about the
HDF5 version, but things seem to work anyway.

TODO: Figure out how to fix this mess.