DAK-2346: Add PDFs to HDF5 output
I grabbed the response labels from the iteratedModel, which
may turn out not to work in some instances (ex. recast models).
Because of how the lower and upper bounds are stored by Dakota,
I added another constructor to RealScale that takes a pointer
to Real and a length. This is to avoid copying data.