DAK-2370: Improve encapsulation of ResultsManager and ResultsDBHDF5
  - created a new base ResultsDBBase base class patterned after ResultsDBAny
  - removed some unused methods
  - commented out other unused methods that might be worth keeping for
    ideas going forward
  - simplified the header/class dependencies to improve compile times when
    modifying source and headers, eg
    * touching ResultsDBAny.[ch]pp or ResultsDBHAF5.[ch]pp recompiles a few
      other files instead of the previous dozens

I left the ResultsManager client API using templates but then use
interitance via boost::any types and the new base class to avoid
the need to include implementation details in either ResultsManager
or ResultsDBBase.  This seems to achieve the goal of this story.
Some more cleanup is definitely in order along with a code review by
Adam and Elliott.