ResultsDB: Enforce unique user-provided block IDs
For DAK-2383: Require any user-specified id_{method, variables, etc.}
strings be unique across blocks of each type in the input file. This
primarily supports unique IDs for methods for HDF5 output, but also
generally enforces best practice.

It allows a variables and response block to have the same ID, but not
two variables blocks, for instance.