Add flush to ResultsManager
To reduce the risk of corruption due to crashes or other
unexpected interruptions, the HDF5 cache should be occasionally
flushed to disk. The current strategy is to flush at the
conclusion of each method and in the abort handler.

The effectiveness and performance consequences of changing the
the cache size file access property should also be investigated.

* Rename ResultsManager::write_databases() and
  ResultsDBAny::print_results() to flush().
* Add flush() to ResultsDBHDF5 and HDF5IOHelper
* Remove call to resultsDB.write_databases() from
  Analyzer::post_run() and Minimizer::post_run() and place a
  call to resultsDB.flush() at the end of Iterator::run().
* Place a call to flush() in the global abort_handler().
* For ResultsDBAny, flush() writes the cached results to the
  text file
* For ResultsDBHDF5/HDF5IOHelper, flush() calls filePtr->flush().