OutputDB: Tag files when running in parallel
For DAK-2436: Create one output database per Dakota process, tagging
when running in MPI mode. (For now, will tag with one-based id, even
if mpirun -n 1; could be improved to suppress tag in that case.)

Adam and I discussed an alternate design where we use hierarchical
tagging with a database per active iterator, and/or only outputting a
database on ranks serving an iterator, but all this may go away later
with MPI-enabled DB output.

The previous implementation was incorrect, using iterator hierarchical
tagging when there is one global ResultsManager, and not performing
the initialization at the right point. The databases should be
initialized and finalized once per execution, not per iterator run as
multiple iterators may write to them.