Fix test results for dakota_sbuq_hier_poly_prod.in
All of the test results for dakota_sbuq_hier_poly_prod.in were consistently
wrong.  This was due to the badScalingFlag logic in DiscrepancyCorrection for
falling back from a multiplicative to an additive correction when division by
zero FPEs are encountered.  This works well for multifidelity optimization
using local corrections but not for multifidelity PCE,SC, which do not support
flipping the correction type on the fly.  So discrepancy calculations were
falling back to being computed additively while the expansions were still
being combined multiplicatively.  Now multifidelity PCE & SC abort with an
error when multiplicative FPEs are encountered.

To fix the test set, the random domain was previously offset away from 0.
This commit updates the baselines with new multiplicative expansion results,
now verified against the single fidelity case.