DAK-2421: Store optimization/calibration results
TODO:
1. Treat configuration/inactive variables consistently with
the console.

Right now, only active variables are stored. Console output
is conditional (and inconsistent between LeastSq and Optimizer).
In LeastSq, when no configuration variables are present, inactive
variables are printed with the best parameters. With config vars,
they are printed only with their corresponding original model
responses. In Optimizer, inactive variables are always printed
with the best parameters, so a "random" configuration variable
may be included.

2. Add ability to store 3D datasets

Arguably the best way to store the original response data
for a calibration with multiple experiments is in 3D datasets.