Blessing #s3 for QUESO using latest ML PCE regression.
After careful review, a rounding change in the ML sample allocation induces
a small change in the sample profile and in the multilevel compressed sensing
emulator.  The MAP point is similar but there are some significant changes in
some chain stats, while finding a solution with lower misfit.  These differences
seem attributable to the short chain length (2000 samples) and lack of posterior
convergence in 9D, even with Hessian pre-conditioning.

This finishes off the review & resolution of serial DIFFs.  Up tomorrow,
resolve a few parallel tests and add new ML tests.