Abort when var or resp descriptors are repeated
Add a parse-time check to "sqawk" when descriptors are
repeated in a response or variables block. This change heads
off an issue with HDF5 results storage, where attempted addition
of multiple datasets of the same name is an error. It's also
just common sense to disambiguate input/output.

A couple of tests had repeated variable descriptors and
had to be updated. Changes to the baseline were to descriptor
labels only, not to numerical results.