Revise order of operations for greedy iso / aniso refinements.
First pop the approximation coefficients and then pop the shared data for
uniform and anisotropic refinement, as previously done for generalized
sparse grid refinement.  This is needed to allow use of shared data when
decrementing / finalizing expansion coefficients within greedy ML PCE/SC.

Note: as indicated in the associated Pecos commit, this moves in the direction
of destabilizing multilevel_harden, which will be resolved at some point in
the future.