Bug: Post-run input fails for freeform data
Fix DAK-1350, where post-run input fails for freeform data files.

Due to no eval ID to import, the IntResponseMap ended up with a single
entry for evalID 0. Now they are numbered starting from 1 on initial
import for the freeform case (and the actual IDs are used for the
annotated case), which allows insertion into the Response map.

This should be compatible with the data import scheme and cache /
restart insertion used in DataFitSurrModel.