Comment out std::vector template specialization for MPI buffers for now.
Parallel tests in the COLINY portion of dakota_hybrid.in are failing when
std::vector is assigned based on resize(len) and operator[], instead of
assigned using push_back().  Not sure why, but a guess would be that memory
is being written out of expected bounds, and the factor of 2 growth from
push_back() provides a cushion that hides the latent bug.