Remove extra escaped quotes from add_application_test args
These throw warnings when the paths to the files contain spaces, and
it seems CMake handles these arguments correctly on Windows, even when
the directory names have spaces.

Opted not to use nested + escaped quotes as the AddApplicationTest
unit tests do as that also exhibits the problem we're trying to fix.