DAK-2527 : Changes to use updated Trilinos tpl
 - uses a trilinos patch to circumvent the new Trilinos requirement
   on cmake version >= 3.10
 - update local sha1
 - update pecos and external sha1's

 - running all tests under the RHEL7 docker container (opt build) shows
   the following non-passing tests:

The following tests FAILED:
         18 - dakota_apps (Failed)
         69 - dakota_nomad (Failed)
        122 - dakota_rol (Failed)
        125 - dakota_rosenbrock_ext (Failed)
        182 - dakota_textbook_lsq (Failed)

Inspection of the corresponding dakota_diffs.out file shows that these
are all small diffs except for the dakota_rosenbrock_ext test which has
non-trivial differences for tests s1, s8 and s15, all of which
use dot_frcg.  I looked a little more into this until I (re)discovered
the blurb in the Reference Manual:

"We here provide a caution regarding dot_frcg. In DOT Version 4.20, we
have noticed inconsistent behavior of this algorithm across different
versions of Linux. Our best assessment is that it is due to different
treatments of uninitialized variables. As we do not know the intention
of the code authors and maintaining DOT source code is outside of the
Dakota project scope, we have not made nor are we recommending any code
changes to address this. However, all users who use dot_frcg in DOT
Version 4.20 should be aware that results may not be reliable."

Apparently, the update to Trilinos now causes this known memory issue
to manifest on the gold rhel7 platform and might need to be addressed
more going into the next release.