Perl test script: don't FAIL on abort/error in stderr
For DAK-2645. Historically for regression tests, stderr was examined
and the presence of abort or error was used to communicate
failure. This helped deal with odd MPI exit behavior, but can yield
false negative FAILs when TPLs emit those strings to stderr.

Disabled the check, as it's a weak test anyway that doesn't get at
what we really want to test.