DAK-1933 : Disable Demo_TPL by default
For now, the code and test associated with the Demo_TPL is optionally
enabled via -DHAVE_DEMO_TPL:BOOL=ON .

I confirmed everything compiles and tests run with and without this
option enabled.  I didn't add any baselines for the dakota_demo_tpl test
because it doesn't do anything at present (so it "passes" for now).