Optimizers with user functions don't store evals
A few optimizers can be constructed with user-defined functions.
For example, NCSU is used (I believe) by Dakota's GP for fitting.

Prevent this iterators from trying to store evaluations.

EvaluationStore now also catches when a model is resized
(Hierarchical surrogates seem to do this..) and reports a more-
or-less helpful warning to the screen instead of just bombing.
This is a regret for this release.