Memory access fix in gerstner_string test driver
The standalone gerstner_string driver was implemented for 2 real and 1
string variable, but the indexing carried over from the direct driver
was overrunning the real array by 1. I generalized most of the driver
to arbitrary number of real variables, but left the error trap for 2
reals in for now.

Updated baselines accordingly and reviewed with @lpswile. The
alternative was to hardwire the 3rd real variable to zero as that's
the lucky initialization we were getting on RHEL7.