Store strings without segfaulting on GCC 5.3
For some reason HDF5 and GCC 5+ don't see eye to eye when it
comes to std::strings. This workaround creates a temporary vector
of const char* for HDF5 to store instead.

Also, added a unit test to protect it, which required also adding
an override of HDF5IOHelper::read_vector() for
std::vector<std::string>.

I also reorganized the HDF5IOHelper class by moving template
definitions out of the class decarlation and moving some
non-template functions to the cpp file.