DAK-2666: Enable C3 in Dakota src
 - got the code to compile but only after commenting out some
   variable/response usage
 - replacd solverTol with conergenceTol

It almost appears that there is a Pecos::SurrogateData API intended for
use with the C3Approximation class, but I couldn't find a corresponding
branch of pecos.  It also appears that there is meant to be more than
one Pecos::SurrogateData in play, but C3Approximation only uses one.

These assumptions and the commented out code to make things compile are
decorated with comments containing "RWH" so that Mike, Alex or Gianluca
can easily review this.