Fix gradient/hessian storage bug
In the probably rare situation that a user specified gradients
or hessians but had no continuous variables, Dakota attempted
to allocate a dataset that was size 0 in one dimension, which
generated an unhandled exception.

The fix was to add an additional check on the number of
continuous variables, instead of just looking at the gradient/
hessian specification.