Bless DIFFs for "option 2" (described in previous commits) for unique discrete LHS.
Both option 1 and option 2 accumulate 2 PASS and 4 DIFF across the tests, and
both sets of DIFFs are small.  Option 2 was selected as the most similar to
previous behavior on devel, prior to discussion/resolution next week.

Note that backfill now occurs even in the case when the number of discrete
combinations is less than the desired number of unique samples.  In this case,
backfill iterations are employed up until no additional sample gain is
encountered.

Serial tests are now clean.  Parallel and unit remain prior to merge.